Fix 14 more compile errors: invalid dereference patterns
Fixed pattern (int)*(int)psVarN << shift -> (int)*psVarN << shift
across multiple functions using fixed-point math routines.

Compile errors reduced from 33 to 19.

+++ src/FUN_0602B1D6.c
@@ -119,7 +119,7 @@ long long FUN_0602b1d6(param_1, param_2, param_3, param_4, param_5)
 
   uVar18 = in_sr & 0xfffffffe;
 
-  uVar6 = (int)*(int)psVar5 << 0xb;
+  uVar6 = (int)*psVar5 << 0xb;
 
   *(unsigned int *)0x0602B304 = uVar6;
