+++ src/FUN_0602B55A.c
@@ -1291,7 +1291,7 @@ LAB_0602b972:
 
       sVar2 = psVar15[2];
 
-      *(int *)0x06094AB0 = (int)*(int)psVar15 << 0xb;
+      *(int *)0x06094AB0 = (int)*psVar15 << 0xb;
 
       *(int *)(puVar5 + 4) = (int)sVar1 << 0xb;
 
@@ -3145,7 +3145,7 @@ LAB_0602b972:
 
     sVar2 = psVar15[2];
 
-    *(int *)0x06094AB0 = (int)*(int)psVar15 << 0xb;
+    *(int *)0x06094AB0 = (int)*psVar15 << 0xb;
 
     *(int *)(puVar5 + 4) = (int)sVar1 << 0xb;
 
@@ -3943,7 +3943,7 @@ LAB_0602b972:
 
     sVar2 = psVar15[2];
 
-    *(int *)0x06094AB0 = (int)*(int)psVar15 << 0xb;
+    *(int *)0x06094AB0 = (int)*psVar15 << 0xb;
 
     *(int *)(puVar5 + 4) = (int)sVar1 << 0xb;
