+++ src/FUN_0602B58E.c
@@ -1024,7 +1024,7 @@ LAB_0602b972:
 
       sVar2 = psVar14[2];
 
-      *(int *)0x06094AB0 = (int)*(int)psVar14 << 0xb;
+      *(int *)0x06094AB0 = (int)*psVar14 << 0xb;
 
       *(int *)(puVar4 + 4) = (int)sVar1 << 0xb;
 
@@ -2880,7 +2880,7 @@ LAB_0602b972:
 
     in_sr = uVar17 & 0xfffffffe;
 
-    *(int *)0x06094AB0 = (int)*(int)psVar14 << 0xb;
+    *(int *)0x06094AB0 = (int)*psVar14 << 0xb;
 
     *(int *)(puVar4 + 4) = (int)sVar1 << 0xb;
 
@@ -3680,7 +3680,7 @@ LAB_0602b972:
 
     sVar2 = psVar14[2];
 
-    *(int *)0x06094AB0 = (int)*(int)psVar14 << 0xb;
+    *(int *)0x06094AB0 = (int)*psVar14 << 0xb;
 
     *(int *)(puVar4 + 4) = (int)sVar1 << 0xb;
 
@@ -4482,7 +4482,7 @@ LAB_0602b972:
 
     sVar2 = psVar14[2];
 
-    *(int *)0x06094AB0 = (int)*(int)psVar14 << 0xb;
+    *(int *)0x06094AB0 = (int)*psVar14 << 0xb;
 
     *(int *)((int)param_1 + 4) = (int)sVar1 << 0xb;
