+++ src/FUN_0602B5A2.c
@@ -764,7 +764,7 @@ LAB_0602b972:
 
       sVar2 = psVar14[2];
 
-      *(int *)0x06094AB0 = (int)*(int)psVar14 << 0xb;
+      *(int *)0x06094AB0 = (int)*psVar14 << 0xb;
 
       *(int *)(puVar4 + 4) = (int)sVar1 << 0xb;
 
@@ -2620,7 +2620,7 @@ LAB_0602b972:
 
     in_sr = uVar17 & 0xfffffffe;
 
-    *(int *)0x06094AB0 = (int)*(int)psVar14 << 0xb;
+    *(int *)0x06094AB0 = (int)*psVar14 << 0xb;
 
     *(int *)(puVar4 + 4) = (int)sVar1 << 0xb;
 
@@ -3420,7 +3420,7 @@ LAB_0602b972:
 
     sVar2 = psVar14[2];
 
-    *(int *)0x06094AB0 = (int)*(int)psVar14 << 0xb;
+    *(int *)0x06094AB0 = (int)*psVar14 << 0xb;
 
     *(int *)(puVar4 + 4) = (int)sVar1 << 0xb;
 
@@ -4222,7 +4222,7 @@ LAB_0602b972:
 
     sVar2 = psVar14[2];
 
-    *(int *)0x06094AB0 = (int)*(int)psVar14 << 0xb;
+    *(int *)0x06094AB0 = (int)*psVar14 << 0xb;
 
     *(int *)((int)param_1 + 4) = (int)sVar1 << 0xb;
