+++ src/FUN_0602B5BE.c
@@ -502,7 +502,7 @@ LAB_0602b972:
 
       sVar2 = psVar15[2];
 
-      *(int *)0x06094AB0 = (int)*(int)psVar15 << 0xb;
+      *(int *)0x06094AB0 = (int)*psVar15 << 0xb;
 
       *(int *)(puVar5 + 4) = (int)sVar1 << 0xb;
 
@@ -2358,7 +2358,7 @@ LAB_0602b972:
 
     in_sr = uVar18 & 0xfffffffe;
 
-    *(int *)0x06094AB0 = (int)*(int)psVar15 << 0xb;
+    *(int *)0x06094AB0 = (int)*psVar15 << 0xb;
 
     *(int *)(puVar5 + 4) = (int)sVar1 << 0xb;
 
@@ -3156,7 +3156,7 @@ LAB_0602b972:
 
     sVar2 = psVar15[2];
 
-    *(int *)0x06094AB0 = (int)*(int)psVar15 << 0xb;
+    *(int *)0x06094AB0 = (int)*psVar15 << 0xb;
 
     *(int *)(puVar5 + 4) = (int)sVar1 << 0xb;
 
@@ -3956,7 +3956,7 @@ LAB_0602b972:
 
     sVar2 = psVar15[2];
 
-    *(int *)0x06094AB0 = (int)*(int)psVar15 << 0xb;
+    *(int *)0x06094AB0 = (int)*psVar15 << 0xb;
 
     *(int *)(puVar5 + 4) = (int)sVar1 << 0xb;
