+++ src/FUN_0602B5D2.c
@@ -498,7 +498,7 @@ LAB_0602b972:
 
       sVar2 = psVar16[2];
 
-      *(int *)0x06094AB0 = (int)*(int)psVar16 << 0xb;
+      *(int *)0x06094AB0 = (int)*psVar16 << 0xb;
 
       *(int *)(puVar6 + 4) = (int)sVar1 << 0xb;
 
@@ -2352,7 +2352,7 @@ LAB_0602b972:
 
     sVar2 = psVar16[2];
 
-    *(int *)0x06094AB0 = (int)*(int)psVar16 << 0xb;
+    *(int *)0x06094AB0 = (int)*psVar16 << 0xb;
 
     *(int *)(puVar7 + 4) = (int)sVar1 << 0xb;
 
@@ -3152,7 +3152,7 @@ LAB_0602b972:
 
     sVar2 = psVar16[2];
 
-    *(int *)0x06094AB0 = (int)*(int)psVar16 << 0xb;
+    *(int *)0x06094AB0 = (int)*psVar16 << 0xb;
 
     *(int *)(puVar7 + 4) = (int)sVar1 << 0xb;
 
@@ -3954,7 +3954,7 @@ LAB_0602b972:
 
     sVar2 = psVar16[2];
 
-    *(int *)0x06094AB0 = (int)*(int)psVar16 << 0xb;
+    *(int *)0x06094AB0 = (int)*psVar16 << 0xb;
 
     *(int *)(puVar7 + 4) = (int)sVar1 << 0xb;
