+++ src/FUN_0602B8DC.c
@@ -164,7 +164,7 @@ LAB_0602b972:
 
   sVar2 = psVar17[2];
 
-  *(int *)0x06094AB0 = (int)*(int)psVar17 << 0xb;
+  *(int *)0x06094AB0 = (int)*psVar17 << 0xb;
 
   *(int *)(puVar7 + 4) = (int)sVar1 << 0xb;
 
@@ -1984,7 +1984,7 @@ LAB_0602b444:
 
     uVar21 = uVar20 & 0xfffffffe;
 
-    *(int *)0x06094AB0 = (int)*(int)psVar17 << 0xb;
+    *(int *)0x06094AB0 = (int)*psVar17 << 0xb;
 
     *(int *)(puVar16 + 4) = (int)sVar1 << 0xb;
 
@@ -2780,7 +2780,7 @@ LAB_0602b444:
 
     sVar2 = psVar17[2];
 
-    *(int *)0x06094AB0 = (int)*(int)psVar17 << 0xb;
+    *(int *)0x06094AB0 = (int)*psVar17 << 0xb;
 
     *(int *)(puVar16 + 4) = (int)sVar1 << 0xb;
 
@@ -3578,7 +3578,7 @@ LAB_0602b444:
 
     sVar2 = psVar17[2];
 
-    *(int *)0x06094AB0 = (int)*(int)psVar17 << 0xb;
+    *(int *)0x06094AB0 = (int)*psVar17 << 0xb;
 
     *(int *)(puVar16 + 4) = (int)sVar1 << 0xb;
