+++ src/FUN_0602BDCC.c
@@ -131,7 +131,7 @@ unsigned int FUN_0602bdcc(param_1, param_2, param_3, param_4)
 
     uVar20 = in_sr & 0xfffffffe;
 
-    *(int *)0x06094F74 = (int)*(int)psVar15 << 0xb;
+    *(int *)0x06094F74 = (int)*psVar15 << 0xb;
 
     *(int *)(puVar5 + 4) = (int)sVar1 << 0xb;
 
@@ -1985,7 +1985,7 @@ LAB_0602bee8:
 
       uVar21 = uVar20 & 0xfffffffe;
 
-      *(int *)0x06094F74 = (int)*(int)psVar15 << 0xb;
+      *(int *)0x06094F74 = (int)*psVar15 << 0xb;
 
       *(int *)(puVar6 + 4) = (int)sVar1 << 0xb;
 
@@ -2785,7 +2785,7 @@ LAB_0602bee8:
 
       sVar2 = psVar15[2];
 
-      *(int *)0x06094F74 = (int)*(int)psVar15 << 0xb;
+      *(int *)0x06094F74 = (int)*psVar15 << 0xb;
 
       *(int *)(puVar6 + 4) = (int)sVar1 << 0xb;
 
@@ -3587,7 +3587,7 @@ LAB_0602bee8:
 
       sVar2 = psVar15[2];
 
-      *(int *)0x06094F74 = (int)*(int)psVar15 << 0xb;
+      *(int *)0x06094F74 = (int)*psVar15 << 0xb;
 
       *(int *)(puVar6 + 4) = (int)sVar1 << 0xb;
